refactor(auth): Phase 5 signing infra split + AddAuthDatastore rename
- Move IRsaKeyProvider + PublishedKey to Domain/Abstractions/ (namespace
  Auth.Service.Domain.Abstractions); move PemFileRsaKeyProvider +
  SigningOptions to Infrastructure/Signing/; add AddSigningInfrastructure
  binding SigningOptions + registering IRsaKeyProvider singleton.
- Rename AddSqlServerDatastore to AddAuthDatastore (returns
  IServiceCollection); Program.cs chains it with AddSigningInfrastructure;
  RegisterTokenService no longer registers signing concerns.
