+++ auth-microservice/Auth.Service/Domain/Tokens/ServiceTokenService.cs
@@ -3,8 +3,8 @@ using System.Security.Claims;
 using System.Security.Cryptography;
 using System.Text;
 using Auth.Service.Domain;
+using Auth.Service.Domain.Abstractions;
 using Auth.Service.Services;
-using Auth.Service.Services.Signing;
 using ECommerce.Shared.Authentication;
 using Microsoft.IdentityModel.Tokens;
