+++ auth-microservice/Auth.Service/Program.cs
@@ -1,5 +1,6 @@
 using Auth.Service.Endpoints;
 using Auth.Service.Infrastructure.Data.EntityFramework;
+using Auth.Service.Infrastructure.Signing;
 using Auth.Service.Services;
 using ECommerce.Shared.HealthChecks;
 using ECommerce.Shared.Observability;
@@ -9,7 +10,8 @@ using ECommerce.Shared.Qa;
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Services.AddSingleton<Microsoft.AspNetCore.Identity.IPasswordHasher<Auth.Service.Domain.User>, Microsoft.AspNetCore.Identity.PasswordHasher<Auth.Service.Domain.User>>();
-builder.Services.AddSqlServerDatastore(builder.Configuration);
+builder.Services.AddAuthDatastore(builder.Configuration)
+                .AddSigningInfrastructure(builder.Configuration);
 builder.Services.RegisterTokenService(builder.Configuration);
 
 builder.AddPlatformObservability("Auth",
