+++ auth-microservice/Auth.Service/Services/TokenStartupExtensions.cs
@@ -1,5 +1,4 @@
 using Auth.Service.Domain.Tokens;
-using Auth.Service.Services.Signing;
 using ECommerce.Shared.Authentication;
 
 namespace Auth.Service.Services;
@@ -12,11 +11,6 @@ public static class TokenStartupExtensions
         var authOptions = new AuthOptions();
         configuration.GetSection(AuthOptions.AuthenticationSectionName).Bind(authOptions);
         services.AddSingleton(authOptions);
-
-        var signingOptions = new SigningOptions();
-        configuration.GetSection(SigningOptions.SectionName).Bind(signingOptions);
-        services.AddSingleton(signingOptions);
-        services.AddSingleton<IRsaKeyProvider, PemFileRsaKeyProvider>();
 
         services.AddScoped<JwtTokenService>();
         services.AddScoped<LoginHandler>();
