+++ auth-microservice/Auth.Tests/JwtTokenServiceTests.cs
@@ -1,8 +1,8 @@
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Cryptography;
 using Auth.Service.Domain;
+using Auth.Service.Domain.Abstractions;
 using Auth.Service.Domain.Tokens;
-using Auth.Service.Services.Signing;
 using ECommerce.Shared.Authentication;
 using Microsoft.IdentityModel.Tokens;
 using NSubstitute;
