+++ auth-microservice/Auth.Tests/ServiceTokenServiceTests.cs
@@ -1,9 +1,9 @@
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Cryptography;
+using Auth.Service.Domain.Abstractions;
 using Auth.Service.Domain.Tokens;
 using Auth.Service.Models;
 using Auth.Service.Services;
-using Auth.Service.Services.Signing;
 using ECommerce.Shared.Authentication;
 using Microsoft.IdentityModel.Tokens;
